W-13669227: refactor
