+++ pom.xml
@@ -23,7 +23,6 @@
         <c3p0Version>0.9.5.5</c3p0Version>
         <commonsCollectionsVersion>3.2.2</commonsCollectionsVersion>
         <commonsDbUtilsVersion>1.7</commonsDbUtilsVersion>
-        <commonsLangVersion>3.12.0</commonsLangVersion>
         <commonsIoVersion>2.13.0</commonsIoVersion>
         <mockito-core.version>4.7.0</mockito-core.version>
         <junit.version>4.13.2</junit.version>
@@ -54,7 +53,7 @@
         <formatterConfigPath>formatter.xml</formatterConfigPath>
 
         <munit.extensions.maven.plugin.version>1.2.0-rc3</munit.extensions.maven.plugin.version>
-        <munit.version>3.1.0-SNAPSHOT</munit.version>
+        <munit.version>3.1.0-rc4</munit.version>
 
         <mtf-tools.version>1.2.0-rc1</mtf-tools.version>
         <skipTests>false</skipTests>
@@ -77,7 +76,7 @@
         <mule.sdk.api.version>0.7.0</mule.sdk.api.version>
         <commons-dbcp2.version>2.9.0</commons-dbcp2.version>
         <mtf.javaopts></mtf.javaopts>
-        <mule.sdk.api.version>0.7.0</mule.sdk.api.version>
+        <awaitility.version>3.1.6</awaitility.version>
     </properties>
 
     <dependencyManagement>
@@ -132,17 +131,11 @@
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-lang3</artifactId>
             <version>${commons.lang.version}</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>com.mchange</groupId>
             <artifactId>c3p0</artifactId>
             <version>${c3p0Version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>${commonsLangVersion}</version>
         </dependency>
         <dependency>
             <groupId>commons-io</groupId>
@@ -271,7 +264,7 @@
         <dependency>
             <groupId>org.awaitility</groupId>
             <artifactId>awaitility</artifactId>
-            <version>3.1.6</version>
+            <version>${awaitility.version}</version>
             <scope>test</scope>
         </dependency>
 
@@ -290,11 +283,6 @@
         <dependency>
             <groupId>com.sun.istack</groupId>
             <artifactId>istack-commons-runtime</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.mule.sdk</groupId>
-            <artifactId>mule-sdk-api</artifactId>
-            <version>${mule.sdk.api.version}</version>
         </dependency>
     </dependencies>
 
@@ -1654,5 +1642,4 @@
             <url>https://repository.mulesoft.org/nexus/content/repositories/public/</url>
         </pluginRepository>
     </pluginRepositories>
-
 </project>
