+++ src/main/java/org/mule/extension/db/internal/lifecycle/DB2ArtifactLifecycleListener.java
@@ -7,7 +7,6 @@
 package org.mule.extension.db.internal.lifecycle;
 
 import static java.lang.String.format;
-import static java.lang.Thread.getAllStackTraces;
 import static java.sql.DriverManager.deregisterDriver;
 import static java.sql.DriverManager.getDrivers;
 import static org.slf4j.LoggerFactory.getLogger;
