+++ src/test/java/org/mule/extension/db/lifecycle/DerbyArtifactLifecycleListenerTestCase.java
@@ -8,9 +8,7 @@ package org.mule.extension.db.lifecycle;
 
 import static java.lang.Thread.getAllStackTraces;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasProperty;
 import static org.hamcrest.Matchers.is;
@@ -21,14 +19,11 @@ import org.mule.extension.db.internal.lifecycle.DerbyArtifactLifecycleListener;
 import org.mule.sdk.api.artifact.lifecycle.ArtifactLifecycleListener;
 
 import java.sql.DriverManager;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 import java.util.stream.Collectors;
 
 import org.hamcrest.Matcher;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -39,27 +34,18 @@ public class DerbyArtifactLifecycleListenerTestCase extends AbstractArtifactLife
   public static final String DRIVER_PACKAGE = "org.apache.derby";
   public static final String DRIVER_THREAD_NAME = "derby.rawStoreDaemon";
 
-  private List<Thread> previousThreads = new ArrayList<>();
-
   @Before
   /* The thread that Derby creates when using the in-memory base, does not contain information of the contextClassloader
   in which it was created, this field is set to null. To test that the cleaning of the resourceReleaser does not delete
   threads that correspond to other instances, I create a connection to a different database before starting the test
   which should not be affected after the cleaning. As this connection is made with the driver loaded by SPI I cannot
   close it in the After Test because I would be unregistering a driver that may be required by another test suite. */
-  public void addOtherDatabaseThreadInTheContext() throws Exception {
+  public void getPreviousThreads() throws Exception {
     DriverManager.getConnection("jdbc:derby:previousDB;create=true;user=me;password=mine");
     await().until(() -> getAllStackTraces().keySet().stream()
         .anyMatch(thread -> thread.getName().startsWith(getDriverThreadName())));
     previousThreads = getAllStackTraces().keySet().stream()
         .filter(thread -> thread.getName().startsWith(getDriverThreadName())).collect(Collectors.toList());
-  }
-
-  @After
-  public void checkTheOtherDatabaseThreadInTheContext() throws Exception {
-    assertThat(getAllStackTraces().keySet().stream()
-        .filter(thread -> thread.getName().startsWith(getDriverThreadName())).collect(Collectors.toList()),
-               containsInAnyOrder(previousThreads.toArray()));
   }
 
   public DerbyArtifactLifecycleListenerTestCase(String groupId, String artifactId, String version) {
