+++ src/test/java/org/mule/extension/db/lifecycle/OracleArtifactLifecycleListenerTestCase.java
@@ -6,10 +6,7 @@
  */
 package org.mule.extension.db.lifecycle;
 
-import static java.lang.Thread.getAllStackTraces;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasProperty;
 import static org.hamcrest.Matchers.is;
@@ -19,15 +16,10 @@ import static org.hamcrest.core.IsNot.not;
 import org.mule.extension.db.internal.lifecycle.OracleArtifactLifecycleListener;
 import org.mule.sdk.api.artifact.lifecycle.ArtifactLifecycleListener;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
 
 import org.hamcrest.Matcher;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
@@ -67,14 +59,6 @@ public class OracleArtifactLifecycleListenerTestCase extends AbstractArtifactLif
     return OracleLeakTriggerer.class;
   }
 
-  private List<Thread> previousThreads = new ArrayList<>();
-
-  @Before
-  public void getPreviousThreads() throws Exception {
-    previousThreads = getAllStackTraces().keySet().stream()
-        .filter(thread -> thread.getName().startsWith(getDriverThreadName())).collect(Collectors.toList());
-  }
-
   protected Matcher<Iterable<? super Thread>> hasDriverThreadMatcher(ClassLoader target, boolean negateMatcher) {
     Matcher<Iterable<? super Thread>> matcher =
         hasItem(
@@ -82,12 +66,5 @@ public class OracleArtifactLifecycleListenerTestCase extends AbstractArtifactLif
                       hasProperty("name", is(getDriverThreadName())),
                       not(isIn(previousThreads))));
     return negateMatcher ? not(matcher) : matcher;
-  }
-
-  @After
-  public void checkPreviousThreads() throws Exception {
-    assertThat(getAllStackTraces().keySet().stream()
-        .filter(thread -> thread.getName().startsWith(getDriverThreadName())).collect(Collectors.toList()),
-               containsInAnyOrder(previousThreads.toArray()));
   }
 }
